+++ lib/ductwork/models/availability.rb
@@ -8,17 +8,10 @@ module Ductwork
     validates :started_at, presence: true
     validates :pipeline_klass, presence: true
 
+    # NOTE: this method is essentially the middleman antipattern, but we keep
+    # it for symmetry with `Ductwork::Advancement#abandon!`
     def abandon!
-      job = execution.job
-
-      Ductwork::Record.transaction do
-        execution.lock!
-        execution.reload
-
-        return if execution.completed_at.present?
-
-        job.execution_crashed!(execution)
-      end
+      execution.job.execution_crashed!(execution)
     end
   end
 end